Rename TOML parser module
