--- extensions/toml/modules/parser/src/main/java/io/helidon/extensions/toml/parser/TomlParseException.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (c) 2026 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.helidon.extensions.toml.parser;
-
-import java.util.Objects;
-
-import io.helidon.common.Api;
-
-/**
- * TOML parse exception.
- */
-@Api.Incubating
-public class TomlParseException extends RuntimeException {
-    /**
-     * Create a new exception.
-     *
-     * @param message message
-     */
-    public TomlParseException(String message) {
-        super(Objects.requireNonNull(message));
-    }
-
-    /**
-     * Create a new exception.
-     *
-     * @param message message
-     * @param cause cause
-     */
-    public TomlParseException(String message, Throwable cause) {
-        super(Objects.requireNonNull(message), Objects.requireNonNull(cause));
-    }
-}
