--- extensions/toml/modules/parser/src/main/java/module-info.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2026 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import io.helidon.common.features.api.Features;
-import io.helidon.common.features.api.HelidonFlavor;
-
-/**
- * TOML parser implementation.
- */
-@Features.Name("TOML Parser")
-@Features.Description("TOML parser")
-@Features.Flavor({HelidonFlavor.SE, HelidonFlavor.MP})
-@Features.Path({"TOML", "Parser"})
-module io.helidon.extensions.toml.parser {
-    requires transitive io.helidon.builder.api;
-
-    requires static io.helidon.common.features.api;
-
-    exports io.helidon.extensions.toml.parser;
-}
