+++ extensions/toml/modules/toml/src/main/java/io/helidon/extensions/toml/TomlArray.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2026 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.extensions.toml;
+
+import java.util.List;
+
+import io.helidon.common.Api;
+
+/**
+ * TOML array value.
+ */
+@Api.Incubating
+public final class TomlArray implements TomlValue {
+    private final List<TomlValue> values;
+
+    TomlArray(List<TomlValue> values) {
+        this.values = List.copyOf(values);
+    }
+
+    /**
+     * Array values.
+     *
+     * @return immutable array values
+     */
+    public List<TomlValue> values() {
+        return values;
+    }
+}
