+++ extensions/toml/modules/toml/src/main/java/io/helidon/extensions/toml/TomlVersionBehavior.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2026 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.extensions.toml;
+
+import io.helidon.common.Api;
+
+/**
+ * TOML version-specific parser behavior.
+ */
+@Api.Incubating
+public enum TomlVersionBehavior {
+    /**
+     * Require TOML v1.0.0 syntax.
+     */
+    V1_0_0,
+
+    /**
+     * Require TOML v1.1.0 syntax.
+     */
+    V1_1_0,
+
+    /**
+     * Accept all supported TOML syntax without rejecting version-specific features.
+     */
+    BEST_EFFORT;
+
+    boolean allowsV11() {
+        return this != V1_0_0;
+    }
+}
